Moved builder.CopySketchFilesToBuildPath into proper location
